+++ example/main.dart
@@ -5,7 +5,11 @@ void main() {
   {
     "name": "John Doe",
     "age": 42,
-    "student": false
+    "student": false,
+    "dogs": [
+      "Fido",
+      "Rex"
+    ]
   }
   ''';
 
@@ -17,4 +21,11 @@ void main() {
 
   // Example of a missing field.
   print(object['email'].stringOrNull()); // null
+
+  // Example of array funsies.
+  final dogs = object['dogs'].array().mapUnmodifiable((e) => e.string());
+
+  // We get a real List<String> here at this point.
+  // ignore: unnecessary_type_check
+  print(dogs is List<String>);
 }